// packages/web-core/src/api/v1/api.ts
export interface ErrorRsp {
  /**
   * HTTP status code of operation
   * @type {number}
   * @memberof ErrorRsp
   */
  httpStatusCode: number;
  /**
   *
   * @type {string}
   * @memberof ErrorRsp
   */
  message: string;
  /**
   *
   * @type {RequestData}
   * @memberof ErrorRsp
   */
  requestData: RequestData;
  /**
   * Runtime in seconds for this request
   * @type {number}
   * @memberof ErrorRsp
   */
  runtime: number;
  /**
   *
   * @type {ErrorRspAllOfError}
   * @memberof ErrorRsp
   */
  error: ErrorRspAllOfError;
}
/**
 *
 * @export
 * @interface ErrorRspAllOfError
 */
export interface ErrorRspAllOfError {
  /**
   * Type of error
   * @type {string}
   * @memberof ErrorRspAllOfError
   */
  type: string;
  /**
   * Details of error
   * @type {string}
   * @memberof ErrorRspAllOfError
   */
  details?: string;
  /**
   * Validation errors per field
   * @type {Array<ErrorRspAllOfErrorValidation>}
   * @memberof ErrorRspAllOfError
   */
  validation?: Array<ErrorRspAllOfErrorValidation>;
  /**
   * Additional links to help understand the error
   * @type {Array<string>}
   * @memberof ErrorRspAllOfError
   */
  links: Array<string>;
}
/**
 *
 * @export
 * @interface ErrorRspAllOfErrorValidation
 */
export interface ErrorRspAllOfErrorValidation {
  /**
   *
   * @type {string}
   * @memberof ErrorRspAllOfErrorValidation
   */
  field: string;
  /**
   *
   * @type {string}
   * @memberof ErrorRspAllOfErrorValidation
   */
  message: string;
}
// ...
export interface RequestData {
  /**
   * Unique ID of request, you can provide your own while making the request, if not the ID will be randomly generated on server side
   * @type {string}
   * @memberof RequestData
   */
  requestID: string;
  /**
   * Link to home with details about request
   * @type {string}
   * @memberof RequestData
   */
  link: string;
}